First correct render produced
My interpretation of the soldermask was wrong, I needed to render the
copper twice (one in bright green for traces covered by the soldermask),
one in gold (for pads), then mask the latter one with the soldermask
layer and then blit it on top of the bright green one. Therefore
introduced numerous changes to the way image composition works to allow
for this. Bunch of other features (all kinds of cariro blitting
operators) implemented as well because I was chasing ghostbugs. Also
implemented the drill rendering parser, that was quite straightforward.